Refactor ChampionshipSlots
* Refactor ChampionshipSlots to Championships context
* See #797
