+++ lib/ex338_web/controllers/championship_slot_admin_controller.ex
@@ -1,10 +1,10 @@
 defmodule Ex338Web.ChampionshipSlotAdminController do
   use Ex338Web, :controller
 
-  alias Ex338.{ChampionshipSlot}
+  alias Ex338.{Championships}
 
   def create(conn, %{"fantasy_league_id" => league_id, "championship_id" => id}) do
-    case ChampionshipSlot.Store.create_slots_for_league(id, league_id) do
+    case Championships.create_slots_for_league(id, league_id) do
       {:ok, _} ->
         conn
         |> put_flash(:info, "Slots successfully created.")
