+++ priv/repo/dev_seeds.exs
@@ -23,9 +23,9 @@ defmodule Ex338.DevSeeds do
     FantasyTeams.FantasyTeam,
     Waivers.Waiver,
     FantasyTeams.Owner,
-    ChampionshipResult,
-    ChampionshipSlot,
-    ChampWithEventsResult,
+    Championships.ChampionshipResult,
+    Championships.ChampionshipSlot,
+    Championships.ChampWithEventsResult,
     CalendarAssistant,
     FantasyLeagues.LeagueSport,
     Trades.Trade,
