--- test/ex338/championship_slot_test.exs
@@ -1,18 +1,0 @@
-defmodule Ex338.ChampionshipSlotTest do
-  use Ex338.DataCase
-
-  alias Ex338.ChampionshipSlot
-
-  @valid_attrs %{slot: 42}
-  @invalid_attrs %{}
-
-  test "changeset with valid attributes" do
-    changeset = ChampionshipSlot.changeset(%ChampionshipSlot{}, @valid_attrs)
-    assert changeset.valid?
-  end
-
-  test "changeset with invalid attributes" do
-    changeset = ChampionshipSlot.changeset(%ChampionshipSlot{}, @invalid_attrs)
-    refute changeset.valid?
-  end
-end
