+++ test/ex338/championships/create_slot_test.exs
@@ -1,6 +1,6 @@
-defmodule Ex338.ChampionshipSlot.CreateSlotTest do
+defmodule Ex338.Championships.CreateSlotTest do
   use Ex338.DataCase
-  alias Ex338.{ChampionshipSlot.CreateSlot, ChampionshipSlot, FantasyTeams.FantasyTeam, Repo}
+  alias Ex338.{Championships.CreateSlot, Championships.ChampionshipSlot, FantasyTeams.FantasyTeam, Repo}
 
   describe "create_slots_from_positions/1" do
     test "creates roster slots from positions" do
